Create class for gpu_smi, move nvidia smi
