+++ lisa/tools/gpu_smi.py
@@ -2,13 +2,62 @@
 # Licensed under the MIT license.
 
 import re
+from abc import abstractmethod
 from typing import List, Type
 
 from lisa.executable import Tool
 from lisa.util import LisaException, find_groups_in_lines
 
 
-class AmdSmi(Tool):
+class GpuSmi(Tool):
+    """
+    Base class for GPU monitoring tools (nvidia-smi, amd-smi, etc.).
+    """
+
+    @abstractmethod
+    def get_gpu_count(self) -> int:
+        """Get the number of GPUs detected by the SMI tool"""
+        raise NotImplementedError
+
+
+class NvidiaSmi(GpuSmi):
+    # tuple of gpu device names and their device id pattern
+    # e.g. Tesla GPU device has device id "47505500-0001-0000-3130-444531303244"
+    # A10-4Q device id "56475055-0002-0000-3130-444532323336"
+    gpu_devices = (
+        ("Tesla", "47505500", 0),
+        ("A100", "44450000", 6),
+        ("H100", "44453233", 0),
+        ("A10-4Q", "56475055", 0),
+        ("A10-8Q", "3e810200", 0),
+        ("GB200", "42333130", 0),
+    )
+
+    @property
+    def command(self) -> str:
+        return "nvidia-smi"
+
+    @property
+    def can_install(self) -> bool:
+        return False
+
+    def get_gpu_count(self) -> int:
+        result = self.run("-L")
+        if result.exit_code != 0 or (result.exit_code == 0 and result.stdout == ""):
+            result = self.run("-L", sudo=True)
+            if result.exit_code != 0 or (result.exit_code == 0 and result.stdout == ""):
+                raise LisaException(
+                    f"nvidia-smi command exited with exit_code {result.exit_code}"
+                )
+        gpu_types = [x[0] for x in self.gpu_devices]
+        device_count = 0
+        for gpu_type in gpu_types:
+            device_count += result.stdout.count(gpu_type)
+
+        return device_count
+
+
+class AmdSmi(GpuSmi):
     # Pattern to match GPU entries in amd-smi list output
     # Example:
     # GPU: 0
@@ -25,7 +74,7 @@ class AmdSmi(Tool):
 
     @property
     def can_install(self) -> bool:
-        return False
+        return True
 
     @property
     def dependencies(self) -> List[Type[Tool]]:
